refactor(typescript): utils.GetArrayBuffer
